refactor to ApplicationForLeaveStatisticsService interface
